Add array_reverse function to datafusion-function-* crate
