+++ datafusion/proto/proto/datafusion.proto
@@ -661,23 +661,23 @@ enum ScalarFunction {
   ArrayIntersect = 119;
   ArrayUnion = 120;
   OverLay = 121;
-  /// 122 is Range
+  // 122 is Range
   ArrayExcept = 123;
   // 124 was ArrayPopFront
   Levenshtein = 125;
   SubstrIndex = 126;
   FindInSet = 127;
-  /// 128 was ArraySort
-  /// 129 was ArrayDistinct
-  /// 130 was ArrayResize
+  // 128 was ArraySort
+  // 129 was ArrayDistinct
+  // 130 was ArrayResize
   EndsWith = 131;
-  /// 132 was InStr
-  /// 133 was MakeDate
-  ArrayReverse = 134;
-  /// 135 is RegexpLike
-  /// 136 was ToChar
-  /// 137 was ToDate
-  /// 138 was ToUnixtime
+  // 132 was InStr
+  // 133 was MakeDate
+  // 134 was ArrayReverse
+  // 135 is RegexpLike
+  // 136 was ToChar
+  // 137 was ToDate
+  // 138 was ToUnixtime
 }
 
 message ScalarFunctionNode {
